Move ping functions
Move send_<ping-method>() methods into ping.c.
Make seperate IPv4 and IPv6 versions of the ping
functions static so that only the high level send
functions are available.
