+++ boreas/alivedetection.c
@@ -97,137 +97,6 @@ sniffer_thread (__attribute__ ((unused)) void *vargp)
 }
 
 /**
- * @brief Is called in g_hash_table_foreach(). Check if ipv6 or ipv4, get
- * correct socket and start appropriate ping function.
- *
- * @param key Ip string.
- * @param value Pointer to gvm_host_t.
- * @param scanner_p Pointer to scanner struct.
- */
-static void
-send_icmp (__attribute__ ((unused)) gpointer key, gpointer value,
-           gpointer scanner_p)
-{
-  struct scanner *scanner;
-  struct in6_addr dst6;
-  struct in6_addr *dst6_p = &dst6;
-  struct in_addr dst4;
-  struct in_addr *dst4_p = &dst4;
-  static int count = 0;
-
-  scanner = (struct scanner *) scanner_p;
-
-  count++;
-  if (count % BURST == 0)
-    usleep (BURST_TIMEOUT);
-
-  if (gvm_host_get_addr6 ((gvm_host_t *) value, dst6_p) < 0)
-    g_warning ("%s: could not get addr6 from gvm_host_t", __func__);
-  if (dst6_p == NULL)
-    {
-      g_warning ("%s: destination address is NULL", __func__);
-      return;
-    }
-  if (IN6_IS_ADDR_V4MAPPED (dst6_p) != 1)
-    {
-      send_icmp_v6 (scanner->icmpv6soc, dst6_p, ICMP6_ECHO_REQUEST);
-    }
-  else
-    {
-      dst4.s_addr = dst6_p->s6_addr32[3];
-      send_icmp_v4 (scanner->icmpv4soc, dst4_p);
-    }
-}
-
-/**
- * @brief Is called in g_hash_table_foreach(). Check if ipv6 or ipv4, get
- * correct socket and start appropriate ping function.
- *
- * @param key Ip string.
- * @param value Pointer to gvm_host_t.
- * @param scanner_p Pointer to scanner struct.
- */
-static void
-send_tcp (__attribute__ ((unused)) gpointer key, gpointer value,
-          gpointer scanner_p)
-{
-  struct scanner *scanner;
-  struct in6_addr dst6;
-  struct in6_addr *dst6_p = &dst6;
-  struct in_addr dst4;
-  struct in_addr *dst4_p = &dst4;
-  static int count = 0;
-
-  scanner = (struct scanner *) scanner_p;
-
-  count++;
-  if (count % BURST == 0)
-    usleep (BURST_TIMEOUT);
-
-  if (gvm_host_get_addr6 ((gvm_host_t *) value, dst6_p) < 0)
-    g_warning ("%s: could not get addr6 from gvm_host_t", __func__);
-  if (dst6_p == NULL)
-    {
-      g_warning ("%s: destination address is NULL", __func__);
-      return;
-    }
-  if (IN6_IS_ADDR_V4MAPPED (dst6_p) != 1)
-    {
-      send_tcp_v6 (scanner, dst6_p);
-    }
-  else
-    {
-      dst4.s_addr = dst6_p->s6_addr32[3];
-      send_tcp_v4 (scanner, dst4_p);
-    }
-}
-
-/**
- * @brief Is called in g_hash_table_foreach(). Check if ipv6 or ipv4, get
- * correct socket and start appropriate ping function.
- *
- * @param key Ip string.
- * @param value Pointer to gvm_host_t.
- * @param scanner_p Pointer to scanner struct.
- */
-static void
-send_arp (__attribute__ ((unused)) gpointer key, gpointer value,
-          gpointer scanner_p)
-{
-  struct scanner *scanner;
-  struct in6_addr dst6;
-  struct in6_addr *dst6_p = &dst6;
-  struct in_addr dst4;
-  struct in_addr *dst4_p = &dst4;
-  static int count = 0;
-
-  scanner = (struct scanner *) scanner_p;
-
-  count++;
-  if (count % BURST == 0)
-    usleep (BURST_TIMEOUT);
-
-  if (gvm_host_get_addr6 ((gvm_host_t *) value, dst6_p) < 0)
-    g_warning ("%s: could not get addr6 from gvm_host_t", __func__);
-  if (dst6_p == NULL)
-    {
-      g_warning ("%s: destination address is NULL", __func__);
-      return;
-    }
-  if (IN6_IS_ADDR_V4MAPPED (dst6_p) != 1)
-    {
-      /* IPv6 does simulate ARP by using the Neighbor Discovery Protocol with
-       * ICMPv6. */
-      send_icmp_v6 (scanner->arpv6soc, dst6_p, ND_NEIGHBOR_SOLICIT);
-    }
-  else
-    {
-      dst4.s_addr = dst6_p->s6_addr32[3];
-      send_arp_v4 (scanner->arpv4soc, dst4_p);
-    }
-}
-
-/**
  * @brief Start up the sniffer thread.
  *
  * @param sniffer_thread_id pthread_t thread id.
